+++ src/main/java/com/climbx/climbx/auth/dto/UserOauth2InfoResponse.java
@@ -3,12 +3,11 @@ package com.climbx.climbx.auth.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
-import java.math.BigInteger;
 import java.time.Instant;
 
 public record UserOauth2InfoResponse(
     @Positive @NotNull
-    BigInteger id,
+    Long id,
 
     @NotBlank
     String nickname,
